+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 ### Bugfixes
 
 ### Enhancements & Refactors
-
+- [#1451](https://github.com/openscope/openscope/issues/1451) - Add ability to specify the radial with the `hold` command
 
 # 6.14.2 (October 17, 2019)
 ### Hotfixes